--- a/paddle/fluid/operators/dropout_op_test.cc
+++ b/paddle/fluid/operators/dropout_op_test.cc
@@ -32,14 +32,16 @@ namespace m = paddle::operators::math;
 
 USE_OP(dropout);
 
+static paddle::framework::DDim dims = {10, 10};
+
 void Compare(f::Scope* scope, const p::DeviceContext& ctx) {
   // init
   auto var = scope->Var("X");
   auto tensor = var->GetMutable<f::LoDTensor>();
-  tensor->Resize({10, 10});
+  tensor->Resize(dims);
 
   std::vector<float> init;
-  for (int64_t i = 0; i < 10 * 10; ++i) {
+  for (int64_t i = 0; i < f::product(dims); ++i) {
     init.push_back(1.0);
   }
 
@@ -48,12 +50,12 @@ void Compare(f::Scope* scope, const p::DeviceContext& ctx) {
   auto place = ctx.GetPlace();
   auto out_var = scope->Var("Out");
   auto out_tensor = out_var->GetMutable<f::LoDTensor>();
-  out_tensor->Resize({10, 10});
+  out_tensor->Resize(dims);
   out_tensor->mutable_data<float>(place);  // allocate
 
   auto mask_var = scope->Var("Mask");
   auto mask_tensor = mask_var->GetMutable<f::LoDTensor>();
-  mask_tensor->Resize({10, 10});
+  mask_tensor->Resize(dims);
   mask_tensor->mutable_data<float>(place);  // allocate
 
   // run
@@ -87,18 +89,17 @@ void Compare(f::Scope* scope, const p::DeviceContext& ctx) {
 // https://github.com/PaddlePaddle/Paddle/issues/9507, I temporarily
 // disable this test to remove the prevention of the merge of
 // unrelated PRs.
-/*
+
 TEST(Dropout, CPUDense) {
   f::Scope scope;
   p::CPUPlace place;
   p::CPUDeviceContext ctx(place);
-  Compare(scope, ctx);
+  Compare(&scope, ctx);
 }
 
 TEST(Dropout, GPUDense) {
   f::Scope scope;
   p::CUDAPlace place;
   p::CUDADeviceContext ctx(place);
-  Compare(scope, ctx);
+  Compare(&scope, ctx);
 }
-*/