fix https://github.com/PaddlePaddle/Paddle/issues/9507
fix https://github.com/guoshengCS/transformer-nist/issues/4
the same data in size of 56*80*512, I got a **1.20 ms -> 100us**, 10x acceleration.

before optimize (ms)
```txt
------------------------->     Profiling Report     <-------------------------

Place: All
Time unit: ms
Sorted by total time in descending order in the same thread

Event                                       Calls       Total       Min.        Max.        Ave.
thread0::mul_grad                           4850        6198.66     0.397312    42.3485     1.27807
thread0::dropout                            2500        3016.29     0.559968    2.28013     1.20652
thread0::mul                                4850        3000.64     0.200704    21.7682     0.618688
thread0::softmax_with_cross_entropy         50          2130.97     36.8415     46.8009     42.6194
thread0::layer_norm_grad                    1500        1916.48     0.832512    1.61075     1.27765
thread0::elementwise_add_grad               3700        1513.65     0.043008    3.03923     0.409095
thread0::softmax_grad                       900         1202.94     0.405504    2.22618     1.3366
thread0::layer_norm                         1500        895.745     0.385024    0.750592    0.597163
thread0::matmul_grad                        1800        786.354     0.280576    0.560128    0.436863
thread0::softmax                            900         777.398     0.256256    1.44301     0.863775
```
after optimize 
```txt
==34607== Profiling application: /paddle/Paddle/build/paddle/fluid/operators/dropout_op_test
==34607== Profiling result:
Time(%)      Time     Calls       Avg       Min       Max  Name
 95.25%  2.0010ms         6  333.49us     928ns  1.0640ms  [CUDA memcpy HtoD]
  4.75%  99.713us         1  99.713us  99.713us  99.713us  void paddle::operators::RandomGenerator<float>(unsigned long, float const *, float const , paddle::operators::RandomGenerator<float>*, float const *)

```